requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
clarifai==2.6.2
colorama==0.4.3
configparser==4.0.2
cryptography==2.8
decorator==4.4.2
EasyProcess==0.2.10
emoji==0.5.4
future==0.18.2
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
instaloader==4.2.10
instapy==0.6.8
isort==4.3.21
jsonschema==2.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
MeaningCloud-python==1.2.1
plyer==1.4.2
protobuf==3.11.3
pycparser==2.20
pylint==2.4.4
python-telegram-bot==12.4.2
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
rope==0.16.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
tqdm==4.43.0
urllib3==1.25.8
webdriverdownloader==1.1.0.3
wrapt==1.11.2
